requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
dj-rest-auth==6.0.0
Django==5.0.7
django-allauth==64.0.0
djangorestframework==3.15.2
idna==3.7
iniconfig==2.0.0
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
pytest==8.3.2
pytest-django==4.8.0
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.2
